Simplify and clean up vue components for tracking code
