+++ vue/src/index.ts
@@ -29,7 +29,6 @@ export { default as ContainerManage } from './Container/ContainerManage.vue';
 export { default as ContainerDashboard } from './ContainerDashboard/ContainerDashboard.vue';
 export { default as ContainerSelector } from './ContainerSelector/ContainerSelector.vue';
 export { default as ManageInstallTagCode } from './ManageInstallTagCode/ManageInstallTagCode.vue';
-export { default as TagmanagerTrackingCode } from './TagmanagerTrackingCode/TagmanagerTrackingCode.vue';
 export { default as TrackingCodeCommon } from './TagmanagerTrackingCode/TrackingCodeCommon.vue';
 export { default as TrackingCodePage } from './TagmanagerTrackingCode/TrackingCodePage.vue';
 export { default as TrackingSPAPage } from './TagmanagerTrackingCode/TrackingSPAPage.vue';
